# pkg/handlers/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "sigma",
        "contact": {
            "name": "sigma",
            "url": "https://github.com/go-sigma/sigma"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.3.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/webhooks/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Get a webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.CommonList"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.WebhookItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Delete a webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            }
        },
...
        "/webhooks/{webhook_id}/logs/{webhook_log_id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Get a webhook log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Webhook id",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Webhook log id",
                        "name": "webhook_log_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.CommonList"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.WebhookLogItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Delete a webhook log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Webhook id",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Webhook log id",
                        "name": "webhook_log_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enums.WebhookAction": {
            "type": "string",
            "enum": [
                "Create",
                "Update",
                "Delete",
                "Add",
                "Remove",
                "Ping",
                "Started",
                "Doing",
                "Finished"
            ],
            "x-enum-varnames": [
                "WebhookActionCreate",
                "WebhookActionUpdate",
                "WebhookActionDelete",
                "WebhookActionAdd",
                "WebhookActionRemove",
                "WebhookActionPing",
                "WebhookActionStarted",
                "WebhookActionDoing",
                "WebhookActionFinished"
            ]
        },
        "enums.WebhookResourceType": {
            "type": "string",
            "enum": [
                "Webhook",
                "Namespace",
                "Repository",
                "Tag",
                "Artifact",
                "Member",
                "DaemonTaskGcRepositoryRule",
                "DaemonTaskGcTagRule",
                "DaemonTaskGcArtifactRule",
                "DaemonTaskGcBlobRule",
                "DaemonTaskGcRepositoryRunner",
                "DaemonTaskGcTagRunner",
                "DaemonTaskGcArtifactRunner",
                "DaemonTaskGcBlobRunner"
            ],
            "x-enum-varnames": [
                "WebhookResourceTypeWebhook",
                "WebhookResourceTypeNamespace",
                "WebhookResourceTypeRepository",
                "WebhookResourceTypeTag",
                "WebhookResourceTypeArtifact",
                "WebhookResourceTypeMember",
                "WebhookResourceTypeDaemonTaskGcRepositoryRule",
                "WebhookResourceTypeDaemonTaskGcTagRule",
                "WebhookResourceTypeDaemonTaskGcArtifactRule",
                "WebhookResourceTypeDaemonTaskGcBlobRule",
                "WebhookResourceTypeDaemonTaskGcRepositoryRunner",
                "WebhookResourceTypeDaemonTaskGcTagRunner",
                "WebhookResourceTypeDaemonTaskGcArtifactRunner",
                "WebhookResourceTypeDaemonTaskGcBlobRunner"
            ]
        },
        "errcode.ErrCode": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code provides a unique, string key, often capitalized with\nunderscores, to identify the error code. This value is used as the\nkeyed value when serializing api errors.",
                    "type": "string",
                    "example": "UNAUTHORIZED"
                },
                "description": {
                    "description": "Description provides a complete account of the errors purpose, suitable\nfor use in documentation.",
                    "type": "string",
                    "example": "The access controller was unable to authenticate the client. Often this will be accompanied by a Www-Authenticate HTTP response header indicating how to authenticate."
                },
                "http_status_code": {
                    "description": "HTTPStatusCode provides the http status code that is associated with\nthis error condition.",
                    "type": "integer",
                    "example": 401
                },
                "title": {
                    "description": "Title is a short, human readable description of the error condition\nincluded in API responses.",
                    "type": "string",
                    "example": "Authentication required"
                }
            }
        },
...
        "types.CommonList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {}
                },
                "total": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "types.WebhookItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "enable": {
                    "type": "boolean",
                    "example": true
                },
                "event_artifact": {
                    "type": "boolean",
                    "example": true
                },
                "event_daemon_task_gc": {
                    "type": "boolean",
                    "example": true
                },
                "event_member": {
                    "type": "boolean",
                    "example": true
                },
                "event_namespace": {
                    "type": "boolean",
                    "example": true
                },
                "event_repository": {
                    "type": "boolean",
                    "example": true
                },
                "event_tag": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "namespace_id": {
                    "type": "integer",
                    "example": 1
                },
                "retry_duration": {
                    "type": "integer",
                    "example": 5
                },
                "retry_times": {
                    "type": "integer",
                    "example": 3
                },
                "secret": {
                    "type": "string",
                    "example": "secret"
                },
                "ssl_verify": {
                    "type": "boolean",
                    "example": true
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "url": {
                    "type": "string",
                    "example": "http://example.com/webhook"
                }
            }
        },
        "types.WebhookLogItem": {
            "type": "object",
            "properties": {
                "action": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.WebhookAction"
                        }
                    ],
                    "example": "action"
                },
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "req_body": {
                    "type": "string",
                    "example": ""
                },
                "req_header": {
                    "type": "string",
                    "example": ""
                },
                "resource_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.WebhookResourceType"
                        }
                    ],
                    "example": "event"
                },
                "resp_body": {
                    "type": "string",
                    "example": ""
                },
                "resp_header": {
                    "type": "string",
                    "example": ""
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}